Rename proxy to corsProxy.
Making room for the authProxy option.
