+++ lib/create-app.js
@@ -7,7 +7,7 @@ const uuid = require('uuid')
 const cors = require('cors')
 const LDP = require('./ldp')
 const LdpMiddleware = require('./ldp-middleware')
-const proxy = require('./handlers/proxy')
+const corsProxy = require('./handlers/cors-proxy')
 const SolidHost = require('./models/solid-host')
 const AccountManager = require('./models/account-manager')
 const vhost = require('vhost')
@@ -54,7 +54,12 @@ function createApp (argv = {}) {
 
   // Adding proxy
   if (argv.proxy) {
-    proxy(app, argv.proxy)
+    console.error('The proxy configuration option has been renamed to corsProxy.')
+    argv.corsProxy = argv.proxy
+    delete argv.proxy
+  }
+  if (argv.corsProxy) {
+    corsProxy(app, argv.corsProxy)
   }
 
   // Options handler
