+++ lib/ldp.js
@@ -75,8 +75,8 @@ class LDP {
       this.skin = true
     }
 
-    if (this.proxy && this.proxy[ 0 ] !== '/') {
-      this.proxy = '/' + this.proxy
+    if (this.corsProxy && this.corsProxy[ 0 ] !== '/') {
+      this.corsProxy = '/' + this.corsProxy
     }
 
     debug.settings('Server URI: ' + this.serverUri)
